example/requirements.txt
django==1.6
django-bearbones==0.0.1dev
django-pipeline==1.3.25
django-medusa==0.2.0
django-polymorphic==0.5.6
djangorestframework==2.3.14
south==1.0
